Restructured Files to hopefully match clean architecture
