+++ src/main/java/ch/hutch79/Application/configManager/ConfigManager.java
@@ -1,4 +1,4 @@
-package ch.hutch79.configManager;
+package ch.hutch79.Application.configManager;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.SerializationFeature;
