+++ src/main/java/ch/hutch79/Application/configManager/ConfigMigrator.java
@@ -1,0 +1,4 @@
+package ch.hutch79.Application.configManager;
+
+public class ConfigMigrator {
+}
