--- src/main/java/ch/hutch79/configManager/ConfigMigrator.java
@@ -1,4 +1,0 @@
-package ch.hutch79.configManager;
-
-public class ConfigMigrator {
-}
